#99: Separate product JSON resources with prices and without prices.
